+++ NEWS.md
@@ -1,3 +1,5 @@
+# Current development version
+
 # CBTF 0.4.0 (2025-07-31)
 
 - Don't crash in `get_exported_functions()` if `getNameSpaceExports()` fails.